refactor: 迁移本地 OTEL 观测目录 --story=1070093903136959970
