+++ dev/otel/mock_agent_metrics.py
